Change the buffer storage to a trait value type
Lets the user implement implement custom storage if needed.

Two implementations are provided:

SafeStorageRef stores a reference to an array and can make use of
the borrow checker.

UnsafeStorageRef stores a pointer and length to an array, this will
bypass the borrow-checker but can be useful if one knows that the
usage is ok, but borrow-checker cannot see it (for instance when using
multiple threads but thread is joined before storage dies)